separated methods and datastructure added tests, WIP
